Add JdkFacade and add impl. for Android and Java SE.
Rename and move classes
