+++ java-se/build.gradle.kts
@@ -7,6 +7,9 @@ dependencies {
     implementation("bisq:security")
     implementation("bisq:application")
     implementation("bisq:evolution")
+    implementation("bisq:evolution")
+    implementation("tor:tor")
+    implementation("network:network")
 
     implementation(libs.typesafe.config)
 }
